create source mapper from list of directories to search for source maps in
