+++ ts/src/profiler.ts
@@ -320,10 +320,10 @@ export class Profiler extends ServiceObject {
    * error level and getting profile type will be retried.
    */
   async start() {
-    if (this.config.sourceMapPaths && !this.config.disableSourceMaps) {
+    if (!this.config.disableSourceMaps) {
       try {
         this.sourceMapper =
-            await createSourceMapper(this.config.sourceMapPaths);
+            await createSourceMapper(this.config.sourceMapSearchPath);
       } catch (err) {
         this.logger.error(
             `Failed to initialize source maps and start profiler: ${err}`);
